LayoutTests/http/tests/media/video-play-stall-before-meta-data.html

<video></video>
<p>Test that stalling very early, while loading meta-data, stops delaying the load event.</p>
<script src=../../media-resources/video-test.js></script>
<script>

    // Expected events.
    waitForEvent("beforeload");
    waitForEvent("loadstart");
    waitForEvent("stalled");
    waitForEvent("suspend");

    // We want to make sure the onload event fires.
    window.onload = function() {
       consoleWrite("EVENT(window.onload)");
       endTest();
    };

    // Load should stall very early in the loading process.
    video.src = "http://127.0.0.1:8000/media/video-load-and-stall.cgi?name=../../../media/content/test.mp4&stallAt=1";

</script>
