LayoutTests/http/tests/media/video-play-stall-before-meta-data-expected.txt


Test that stalling very early, while loading meta-data, stops delaying the load event.

EVENT(loadstart)
EVENT(stalled)
EVENT(window.onload)
END OF TEST
